add functionality to surround conflicts in nDRFs to keep other nDRFs enclave
This update will surround conflicts in one-instruction nDRFs,
clearing up any problems with conflicts.
The run pass script has also been changed to add makrings
with the new feature in addition to the existing functionality.

The update appears to work for marking the ll-files,
however there are some shortcommings that myst be considered if using
the xDRF pass as an analysis for other passes.
The nDRFs created to fix conflicts are not
 - inluded as nDRFs in the xDRF structure,
 - sensitive to preceding/following instructions/regions
   (the corresponding sets are empty in the new nDRFs
   and the pre-existing are unchanged),
 - represented in the set of instructions that are in nDRF regions.