change to xchg markings for resndrfs
This commits disables blending of the special resolution nDRFs created by the
CRA feature with the normal nDRFs in the xDRF pass.  Instead they are handled
separately in the marking pass where metadata are attached to the relevant
instructions.  After all traces has been run the runPass script will translate
the metadata into assembly '.ascii' directives with a comment identifying the
beginning or end of a resolution nDRF with its trace number.  After compilation
to assembly code another script, resndrfReplaceXchg, will replace all mentioned
'.ascii' directives with a 'mov' and 'xchg' instruction which constitutes the
actual mark for resolution nDRFs in this revision.